Fix type mismatch

+++ Source/Core/Gen5/States/HiddenGrottoState.hpp
@@ -33,7 +33,7 @@ public:
     {
     }
 
-    u64 getSeed() const
+    u32 getSeed() const
     {
         return seed;
     }
